+++ Modules/Intent.Modules.AspNetCore.OutputCaching.Redis/FactoryExtensions/ControllerExtension.cs
@@ -7,6 +7,7 @@ using Intent.Modelers.Services.Api;
 using Intent.Modelers.Services.CQRS.Api;
 using Intent.Modules.AspNetCore.OutputCaching.Redis.Api;
 using Intent.Modules.Common;
+using Intent.Modules.Common.CSharp.Builder;
 using Intent.Modules.Common.CSharp.Templates;
 using Intent.Modules.Common.Plugins;
 using Intent.Modules.Constants;
@@ -25,7 +26,7 @@ namespace Intent.Modules.AspNetCore.OutputCaching.Redis.FactoryExtensions
         public override string Id => "Intent.AspNetCore.OutputCaching.Redis.ControllerExtension";
 
         [IntentManaged(Mode.Ignore)]
-        public override int Order => 0;
+        public override int Order => 99999999;
 
         protected override void OnAfterTemplateRegistrations(IApplication application)
         {
@@ -33,25 +34,27 @@ namespace Intent.Modules.AspNetCore.OutputCaching.Redis.FactoryExtensions
                 .SelectMany(x => x.Operations.Where(q => q.HasStereotype("Http Settings") && q.GetStereotypeProperty<string>("Http Settings", "Verb") == "GET")).Select(o => o.InternalElement)
                 .ToArray();
 
-            var alloperations = serviceOperations
+            var queryOperations = serviceOperations
                 .Concat(application.MetadataManager.Services(application).GetQueryModels()
                     .Where(x => x.HasStereotype("Http Settings"))
                     .Select(x => x.InternalElement));
 
-            var controllers = new Dictionary<string, List<(IElement Element, CachingAggregate CachingConfig)>>();
-            foreach (var operation in alloperations)
+            var cacheControllers = new Dictionary<string, List<(IElement Element, CachingAggregate CachingConfig)>>();
+            foreach (var operation in queryOperations)
             {
+                // handle caching
                 if (operation.TryGetCaching(out var cachingSettings))
                 {
-                    if (!controllers.TryGetValue(operation.ParentElement.Id, out var operations))
+                    if (!cacheControllers.TryGetValue(operation.ParentElement.Id, out var operations))
                     {
                         operations = new List<(IElement, CachingAggregate)>();
-                        controllers.Add(operation.ParentElement.Id, operations);
+                        cacheControllers.Add(operation.ParentElement.Id, operations);
                     }
                     operations.Add(new(operation, cachingSettings));
                 }
             }
-            foreach (var controller in controllers)
+
+            foreach (var controller in cacheControllers)
             {
                 var template = application.FindTemplateInstance<ICSharpFileBuilderTemplate>(TemplateRoles.Distribution.WebApi.Controller, controller.Key);
                 var endpoints = controller.Value;
@@ -93,6 +96,74 @@ namespace Intent.Modules.AspNetCore.OutputCaching.Redis.FactoryExtensions
                                 att.AddArgument($"VaryByHeaderNames = [{value}]");
                             }
                         });
+                    }
+                });
+            }
+
+            var evictionServiceOperations = application.MetadataManager.Services(application).GetServiceModels()
+                .SelectMany(x => x.Operations.Where(q => q.HasStereotype("Http Settings"))).Select(o => o.InternalElement)
+                .ToArray();
+
+            var allOperation = evictionServiceOperations
+                .Concat(application.MetadataManager.Services(application).GetCommandModels()
+                    .Where(x => x.HasStereotype("Http Settings"))
+                    .Select(x => x.InternalElement));
+
+            var evictionControllers = new Dictionary<string, List<(IElement Element, CacheEvictionAggregate EvictionConfig)>>();
+            foreach (var operation in allOperation)
+            {
+                // handle cache eviction
+                if (operation.TryGetCacheEviction(out var cacheEvictionSettings))
+                {
+                    if (!evictionControllers.TryGetValue(operation.ParentElement.Id, out var operations))
+                    {
+                        operations = [];
+                        evictionControllers.Add(operation.ParentElement.Id, operations);
+                    }
+                    operations.Add(new(operation, cacheEvictionSettings));
+                }
+            }
+            
+            foreach (var controller in evictionControllers)
+            {
+                var template = application.FindTemplateInstance<ICSharpFileBuilderTemplate>(TemplateRoles.Distribution.WebApi.Controller, controller.Key);
+                var endpoints = controller.Value;
+                template.CSharpFile.AfterBuild(file =>
+                {
+                    var @class = file.Classes.FirstOrDefault();
+                    var ctor = @class.Constructors.First();
+
+                    file.AddUsing("Microsoft.AspNetCore.OutputCaching");
+                    if (!ctor.Parameters.Any(p => p.Name == "outputCacheStore"))
+                    {
+                        ctor.AddParameter(file.Template.UseType("Microsoft.AspNetCore.OutputCaching.IOutputCacheStore"), "outputCacheStore", param =>
+                        {
+                            param.IntroduceReadonlyField();
+                        });
+                    }
+
+                    foreach (var endpoint in endpoints)
+                    {
+                        file.AddUsing("Microsoft.AspNetCore.OutputCaching");
+                        var method = @class.FindMethod(m => m.HasMetadata("modelId") && m.GetMetadata<string>("modelId") == endpoint.Element.Id);
+
+                        var index = method.Statements
+                            .Select((statement, idx) => new { statement, idx })
+                            .LastOrDefault(x => x.statement is CSharpReturnStatement)?.idx ?? method.Statements.Count;
+
+                        var insertIndex = (index - 1) > 0 ? index - 1: index;
+
+                        foreach(var tag in endpoint.EvictionConfig?.Tags?.Split(",") ?? [])
+                        {
+                            if(tag.Contains('{'))
+                            {
+                                method.InsertStatement(insertIndex, $"await _outputCacheStore.EvictByTagAsync($\"{tag}\", cancellationToken);");
+                            }
+                            else
+                            {
+                                method.InsertStatement(insertIndex, $"await _outputCacheStore.EvictByTagAsync(\"{tag}\", cancellationToken);");
+                            }
+                        }
                     }
                 });
             }