+++ Modules/Intent.Modules.AspNetCore.OutputCaching.Redis/release-notes.md
@@ -1,5 +1,6 @@
 ### Version 1.1.13
 
+- Improvement: Added the option to evict cache by tag on an endpoint.
 - Improvement: Updated NuGet package versions.
 
 ### Version 1.1.12